+++ sdk/core/Azure.Core/tests/HttpPipelineBuilderTest.cs
@@ -182,7 +182,7 @@ namespace Azure.Core.Tests
             policy.Setup(p => p.OnSendingRequest(It.IsAny<HttpMessage>()))
                 .Callback<HttpMessage>(message =>
                 {
-                    Assert.AreEqual("ExternalClientId",message.Request.ClientRequestId);
+                    Assert.AreEqual("ExternalClientId", message.Request.ClientRequestId);
                     Assert.True(message.Request.TryGetHeader("x-ms-client-request-id", out string requestId));
                     Assert.AreEqual("ExternalClientId", requestId);
                 }).Verifiable();
@@ -275,6 +275,51 @@ namespace Azure.Core.Tests
             else
             {
                 Assert.That(transportField, Is.Not.TypeOf<MockTransport>());
+            }
+        }
+
+        [Test]
+        public async Task TransportOptionsIsClientRedirectEnabledIsOverriddenByClientOptions(
+            [Values(true, false, null)] bool? clientOptionsIsClientRedirectEnabled,
+            [Values(true, false, null)] bool? transportOptionsIsClientRedirectEnabled)
+        {
+            using var testListener = new TestEventListener();
+            testListener.EnableEvents(AzureCoreEventSource.Singleton, EventLevel.Verbose);
+
+            var transport = new MockTransport(
+                new MockResponse(300).AddHeader("Location", "https://new.host/"),
+                new MockResponse(200));
+
+            var options = new TestOptions();
+            if (clientOptionsIsClientRedirectEnabled.HasValue)
+            {
+                options.ClientRedirects = new() { IsClientRedirectEnabled = clientOptionsIsClientRedirectEnabled.Value };
+            }
+            options.Transport = transport;
+
+            var pipeline = HttpPipelineBuilder.Build(new HttpPipelineOptions(options)
+            {
+                ResponseClassifier = ResponseClassifier.Shared
+            }, transportOptionsIsClientRedirectEnabled.HasValue ?
+                new HttpPipelineTransportOptions() { IsClientRedirectEnabled = transportOptionsIsClientRedirectEnabled.Value } :
+                new HttpPipelineTransportOptions());
+
+            using (Request request = pipeline.CreateRequest())
+            {
+                request.Method = RequestMethod.Get;
+                request.Uri.Reset(new Uri("http://example.com"));
+                var response = await pipeline.SendRequestAsync(request, CancellationToken.None);
+
+                if ((clientOptionsIsClientRedirectEnabled.HasValue && clientOptionsIsClientRedirectEnabled.Value) || (transportOptionsIsClientRedirectEnabled ?? false))
+                {
+                    Assert.AreEqual(200, response.Status);
+                    Assert.AreEqual(2, transport.Requests.Count);
+                }
+                else
+                {
+                    Assert.AreEqual(300, response.Status);
+                    Assert.AreEqual(1, transport.Requests.Count);
+                }
             }
         }